requirements.txt
aioboto3==13.2.0
aiobotocore==2.15.2
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aioitertools==0.12.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
asyncio==3.4.3
attrs==24.2.0
boto3==1.35.36
botocore==1.35.36
bson==0.5.10
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
coloredlogs==15.0.1
cryptography==43.0.3
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.7.0
docx2txt==0.8
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.4
flatbuffers==24.3.25
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
httpx-sse==0.4.0
humanfriendly==10.0
idna==3.10
jiter==0.7.0
jmespath==1.0.1
jose==1.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jwt==1.3.1
langchain==0.3.6
langchain-community==0.3.4
langchain-core==0.3.15
langchain-openai==0.2.5
langchain-text-splitters==0.3.1
langsmith==0.1.139
loguru==0.7.2
marshmallow==3.23.1
mongoengine==0.29.1
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
onnxruntime==1.20.0
openai==1.53.0
opencv-python==4.10.0.84
orjson==3.10.10
packaging==24.1
pdf2image==1.17.0
pillow==11.0.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
propcache==0.2.0
protobuf==5.28.3
pyasn1==0.6.1
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
PyJWT==2.9.0
pymongo==4.10.1
pyOpenSSL==24.2.1
pypdf==5.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.17
PyYAML==6.0.2
rapidocr-onnxruntime==1.3.25
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
s3transfer==0.10.3
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
sympy==1.13.3
tenacity==9.0.0
tiktoken==0.8.0
tqdm==4.66.6
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
wrapt==1.16.0
yarl==1.17.1
